% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterMetrics.yarnMetrics}
\alias{ClusterMetrics.yarnMetrics}
\title{ClusterMetrics.yarnMetrics Object}
\usage{
ClusterMetrics.yarnMetrics()
}
\value{
ClusterMetrics.yarnMetrics object
}
\description{
ClusterMetrics.yarnMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The YARN metrics.
}
\seealso{
Other ClusterMetrics functions: \code{\link{ClusterMetrics.hdfsMetrics}},
  \code{\link{ClusterMetrics}}
}
