% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterOperationMetadata}
\alias{ClusterOperationMetadata}
\title{ClusterOperationMetadata Object}
\usage{
ClusterOperationMetadata(ClusterOperationMetadata.labels = NULL,
  operationType = NULL, description = NULL, warnings = NULL,
  labels = NULL, status = NULL, statusHistory = NULL,
  clusterUuid = NULL, clusterName = NULL)
}
\arguments{
\item{ClusterOperationMetadata.labels}{The \link{ClusterOperationMetadata.labels} object or list of objects}

\item{operationType}{Output-only The operation type}

\item{description}{Output-only Short description of operation}

\item{warnings}{Output-only Errors encountered during operation execution}

\item{labels}{Output-only Labels associated with the operation}

\item{status}{Output-only Current operation status}

\item{statusHistory}{Output-only The previous operation status}

\item{clusterUuid}{Output-only Cluster UUID for the operation}

\item{clusterName}{Output-only Name of the cluster for the operation}
}
\value{
ClusterOperationMetadata object
}
\description{
ClusterOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the operation.
}
\seealso{
Other ClusterOperationMetadata functions: \code{\link{ClusterOperationMetadata.labels}}
}
