% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{DiagnoseClusterRequest}
\alias{DiagnoseClusterRequest}
\title{DiagnoseClusterRequest Object}
\usage{
DiagnoseClusterRequest()
}
\value{
DiagnoseClusterRequest object
}
\description{
DiagnoseClusterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to collect cluster diagnostic information.
}
\seealso{
Other DiagnoseClusterRequest functions: \code{\link{projects.regions.clusters.diagnose}}
}

