% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfig}
\alias{GceClusterConfig}
\title{GceClusterConfig Object}
\usage{
GceClusterConfig(GceClusterConfig.metadata = NULL, zoneUri = NULL,
  networkUri = NULL, subnetworkUri = NULL, serviceAccountScopes = NULL,
  tags = NULL, metadata = NULL)
}
\arguments{
\item{GceClusterConfig.metadata}{The \link{GceClusterConfig.metadata} object or list of objects}

\item{zoneUri}{[Required] The zone where the Google Compute Engine cluster will be located}

\item{networkUri}{The Google Compute Engine network to be used for machine communications}

\item{subnetworkUri}{The Google Compute Engine subnetwork to be used for machine communications}

\item{serviceAccountScopes}{The URIs of service account scopes to be included in Google Compute Engine instances}

\item{tags}{The Google Compute Engine tags to add to all instances}

\item{metadata}{The Google Compute Engine metadata entries to add to all instances}
}
\value{
GceClusterConfig object
}
\description{
GceClusterConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common config settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.
}
\seealso{
Other GceClusterConfig functions: \code{\link{GceClusterConfig.metadata}}
}

