% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HadoopJob}
\alias{HadoopJob}
\title{HadoopJob Object}
\usage{
HadoopJob(HadoopJob.properties = NULL, mainJarFileUri = NULL,
  jarFileUris = NULL, loggingConfig = NULL, properties = NULL,
  args = NULL, fileUris = NULL, mainClass = NULL, archiveUris = NULL)
}
\arguments{
\item{HadoopJob.properties}{The \link{HadoopJob.properties} object or list of objects}

\item{mainJarFileUri}{The HCFS URI of the jar file containing the main class}

\item{jarFileUris}{Optional Jar file URIs to add to the CLASSPATHs of the Hadoop driver and tasks}

\item{loggingConfig}{Optional The runtime log config for job execution}

\item{properties}{Optional A mapping of property names to values, used to configure Hadoop}

\item{args}{Optional The arguments to pass to the driver}

\item{fileUris}{Optional HCFS (Hadoop Compatible Filesystem) URIs of files to be copied to the working directory of Hadoop drivers and distributed tasks}

\item{mainClass}{The name of the driver's main class}

\item{archiveUris}{Optional HCFS URIs of archives to be extracted in the working directory of Hadoop drivers and tasks}
}
\value{
HadoopJob object
}
\description{
HadoopJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Dataproc job for running Apache Hadoop MapReduce (https://hadoop.apache.org/docs/current/hadoop-mapreduce-client/hadoop-mapreduce-client-core/MapReduceTutorial.html) jobs on Apache Hadoop YARN (https://hadoop.apache.org/docs/r2.7.1/hadoop-yarn/hadoop-yarn-site/YARN.html).
}
\seealso{
Other HadoopJob functions: \code{\link{HadoopJob.properties}}
}
