% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{InstanceGroupConfig}
\alias{InstanceGroupConfig}
\title{InstanceGroupConfig Object}
\usage{
InstanceGroupConfig(diskConfig = NULL, managedGroupConfig = NULL,
  isPreemptible = NULL, imageUri = NULL, machineTypeUri = NULL,
  instanceNames = NULL, accelerators = NULL, numInstances = NULL)
}
\arguments{
\item{diskConfig}{Optional Disk option config settings}

\item{managedGroupConfig}{Output-only The config for Google Compute Engine Instance Group Manager that manages this group}

\item{isPreemptible}{Optional Specifies that this instance group contains preemptible instances}

\item{imageUri}{Output-only The Google Compute Engine image resource used for cluster instances}

\item{machineTypeUri}{Required The Google Compute Engine machine type used for cluster instances}

\item{instanceNames}{Optional The list of instance names}

\item{accelerators}{Optional The Google Compute Engine accelerator configuration for these instances}

\item{numInstances}{Required The number of VM instances in the instance group}
}
\value{
InstanceGroupConfig object
}
\description{
InstanceGroupConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional The config settings for Google Compute Engine resources in an instance group, such as a master or worker group.
}
