% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Job.labels}
\alias{Job.labels}
\title{Job.labels Object}
\usage{
Job.labels()
}
\value{
Job.labels object
}
\description{
Job.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional The labels to associate with this job. Label keys must contain 1 to 63 characters, and must conform to RFC 1035 (https://www.ietf.org/rfc/rfc1035.txt). Label values may be empty, but, if present, must contain 1 to 63 characters, and must conform to RFC 1035 (https://www.ietf.org/rfc/rfc1035.txt). No more than 32 labels can be associated with a job.
}
\seealso{
Other Job functions: \code{\link{Job}},
  \code{\link{projects.regions.jobs.patch}}
}
