% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{JobReference}
\alias{JobReference}
\title{JobReference Object}
\usage{
JobReference(projectId = NULL, jobId = NULL)
}
\arguments{
\item{projectId}{Required The ID of the Google Cloud Platform project that the job belongs to}

\item{jobId}{Optional The job ID, which must be unique within the project}
}
\value{
JobReference object
}
\description{
JobReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates the full scoping used to reference a job.
}
