% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ListClustersResponse}
\alias{ListClustersResponse}
\title{ListClustersResponse Object}
\usage{
ListClustersResponse(clusters = NULL, nextPageToken = NULL)
}
\arguments{
\item{clusters}{[Output-only] The clusters in the project}

\item{nextPageToken}{[Optional] This token is included in the response if there are more results to fetch}
}
\value{
ListClustersResponse object
}
\description{
ListClustersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The list of all clusters in a project.
}

