% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ListJobsResponse}
\alias{ListJobsResponse}
\title{ListJobsResponse Object}
\usage{
ListJobsResponse(jobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobs}{[Output-only] Jobs list}

\item{nextPageToken}{[Optional] This token is included in the response if there are more results to fetch}
}
\value{
ListJobsResponse object
}
\description{
ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of jobs in a project.
}

