% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{LoggingConfig.driverLogLevels}
\alias{LoggingConfig.driverLogLevels}
\title{LoggingConfig.driverLogLevels Object}
\usage{
LoggingConfig.driverLogLevels()
}
\value{
LoggingConfig.driverLogLevels object
}
\description{
LoggingConfig.driverLogLevels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The per-package log levels for the driver. This may include 'root' package name to configure rootLogger. Examples:  'com.google = FATAL', 'root = INFO', 'org.apache = DEBUG'
}
\seealso{
Other LoggingConfig functions: \code{\link{LoggingConfig}}
}
