% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{LoggingConfig}
\alias{LoggingConfig}
\title{LoggingConfig Object}
\usage{
LoggingConfig(LoggingConfig.driverLogLevels = NULL, driverLogLevels = NULL)
}
\arguments{
\item{LoggingConfig.driverLogLevels}{The \link{LoggingConfig.driverLogLevels} object or list of objects}

\item{driverLogLevels}{The per-package log levels for the driver}
}
\value{
LoggingConfig object
}
\description{
LoggingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The runtime logging config of the job.
}
\seealso{
Other LoggingConfig functions: \code{\link{LoggingConfig.driverLogLevels}}
}
