% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ManagedGroupConfig}
\alias{ManagedGroupConfig}
\title{ManagedGroupConfig Object}
\usage{
ManagedGroupConfig(instanceGroupManagerName = NULL,
  instanceTemplateName = NULL)
}
\arguments{
\item{instanceGroupManagerName}{Output-only The name of the Instance Group Manager for this group}

\item{instanceTemplateName}{Output-only The name of the Instance Template used for the Managed Instance Group}
}
\value{
ManagedGroupConfig object
}
\description{
ManagedGroupConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the resources used to actively manage an instance group.
}
