% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(operationType = NULL, description = NULL, status = NULL,
  state = NULL, details = NULL, clusterName = NULL, clusterUuid = NULL,
  innerState = NULL, endTime = NULL, startTime = NULL, warnings = NULL,
  insertTime = NULL, statusHistory = NULL)
}
\arguments{
\item{operationType}{Output-only The operation type}

\item{description}{Output-only Short description of operation}

\item{status}{Output-only Current operation status}

\item{state}{A message containing the operation state}

\item{details}{A message containing any operation metadata details}

\item{clusterName}{Name of the cluster for the operation}

\item{clusterUuid}{Cluster UUId for the operation}

\item{innerState}{A message containing the detailed operation state}

\item{endTime}{The time that the operation completed}

\item{startTime}{The time that the operation was started by the server}

\item{warnings}{Output-only Errors encountered during operation execution}

\item{insertTime}{The time that the operation was requested}

\item{statusHistory}{Output-only Previous operation status}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the operation.
}
