% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.regions.clusters.delete}
\alias{projects.regions.clusters.delete}
\title{Deletes a cluster in a project.}
\usage{
projects.regions.clusters.delete(projectId, region, clusterName)
}
\arguments{
\item{projectId}{Required The ID of the Google Cloud Platform project that the cluster belongs to}

\item{region}{Required The Cloud Dataproc region in which to handle the request}

\item{clusterName}{Required The cluster name}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}
}
