% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.regions.jobs.cancel}
\alias{projects.regions.jobs.cancel}
\title{Starts a job cancellation request. To access the job resource after cancellation, call regions/{region}/jobs.list or regions/{region}/jobs.get.}
\usage{
projects.regions.jobs.cancel(CancelJobRequest, projectId, region, jobId)
}
\arguments{
\item{CancelJobRequest}{The \link{CancelJobRequest} object to pass to this method}

\item{projectId}{Required The ID of the Google Cloud Platform project that the job belongs to}

\item{region}{Required The Cloud Dataproc region in which to handle the request}

\item{jobId}{Required The job ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}

Other CancelJobRequest functions: \code{\link{CancelJobRequest}}
}
