% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.regions.jobs.list}
\alias{projects.regions.jobs.list}
\title{Lists regions/{region}/jobs in a project.}
\usage{
projects.regions.jobs.list(projectId, region, filter = NULL,
  jobStateMatcher = NULL, pageToken = NULL, pageSize = NULL,
  clusterName = NULL)
}
\arguments{
\item{projectId}{Required The ID of the Google Cloud Platform project that the job belongs to}

\item{region}{Required The Cloud Dataproc region in which to handle the request}

\item{filter}{Optional A filter constraining the jobs to list}

\item{jobStateMatcher}{Optional Specifies enumerated categories of jobs to list (default = match ALL jobs)}

\item{pageToken}{Optional The page token, returned by a previous call, to request the next page of results}

\item{pageSize}{Optional The number of results to return in each response}

\item{clusterName}{Optional If set, the returned jobs list includes only jobs that were submitted to the named cluster}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}
}
