% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SoftwareConfig}
\alias{SoftwareConfig}
\title{SoftwareConfig Object}
\usage{
SoftwareConfig(SoftwareConfig.properties = NULL, imageVersion = NULL,
  properties = NULL)
}
\arguments{
\item{SoftwareConfig.properties}{The \link{SoftwareConfig.properties} object or list of objects}

\item{imageVersion}{Optional The version of software inside the cluster}

\item{properties}{Optional The properties to set on daemon config files}
}
\value{
SoftwareConfig object
}
\description{
SoftwareConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the selection and config of software inside the cluster.
}
\seealso{
Other SoftwareConfig functions: \code{\link{SoftwareConfig.properties}}
}
