% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SparkJob.properties}
\alias{SparkJob.properties}
\title{SparkJob.properties Object}
\usage{
SparkJob.properties()
}
\value{
SparkJob.properties object
}
\description{
SparkJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] A mapping of property names to values, used to configure Spark. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/spark/conf/spark-defaults.conf and classes in user code.
}
\seealso{
Other SparkJob functions: \code{\link{SparkJob}}
}

