% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SparkSqlJob.scriptVariables}
\alias{SparkSqlJob.scriptVariables}
\title{SparkSqlJob.scriptVariables Object}
\usage{
SparkSqlJob.scriptVariables()
}
\value{
SparkSqlJob.scriptVariables object
}
\description{
SparkSqlJob.scriptVariables Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional Mapping of query variable names to values (equivalent to the Spark SQL command: SET name='value';).
}
\seealso{
Other SparkSqlJob functions: \code{\link{SparkSqlJob.properties}},
  \code{\link{SparkSqlJob}}
}
