% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SubmitJobRequest}
\alias{SubmitJobRequest}
\title{SubmitJobRequest Object}
\usage{
SubmitJobRequest(job = NULL)
}
\arguments{
\item{job}{[Required] The job resource}
}
\value{
SubmitJobRequest object
}
\description{
SubmitJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to submit a job.
}
\seealso{
Other SubmitJobRequest functions: \code{\link{projects.regions.jobs.submit}}
}

