% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Cluster}
\alias{Cluster}
\title{Google Cloud Dataproc API Objects 
Manages Hadoop-based clusters and jobs on Google Cloud Platform.}
\usage{
Cluster(projectId = NULL, clusterName = NULL, config = NULL,
  status = NULL, statusHistory = NULL, clusterUuid = NULL)
}
\arguments{
\item{projectId}{[Required] The Google Cloud Platform project ID that the cluster belongs to}

\item{clusterName}{[Required] The cluster name}

\item{config}{[Required] The cluster config}

\item{status}{[Output-only] Cluster status}

\item{statusHistory}{[Output-only] The previous cluster status}

\item{clusterUuid}{[Output-only] A cluster UUID (Unique Universal Identifier)}
}
\value{
Cluster object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:18:50
filename: /Users/mark/dev/R/autoGoogleAPI/googledataprocv1.auto/R/dataproc_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Cluster Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the identifying information, config, and status of a cluster of Google Compute Engine instances.
}
\seealso{
Other Cluster functions: \code{\link{projects.regions.clusters.create}},
  \code{\link{projects.regions.clusters.patch}}
}

