% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterOperationMetadata}
\alias{ClusterOperationMetadata}
\title{ClusterOperationMetadata Object}
\usage{
ClusterOperationMetadata(clusterName = NULL, clusterUuid = NULL,
  status = NULL, statusHistory = NULL, operationType = NULL,
  description = NULL)
}
\arguments{
\item{clusterName}{Name of the cluster for the operation}

\item{clusterUuid}{Cluster UUId for the operation}

\item{status}{[Output-only] Current operation status}

\item{statusHistory}{[Output-only] The previous operation status}

\item{operationType}{[Output-only] The operation type}

\item{description}{[Output-only] Short description of operation}
}
\value{
ClusterOperationMetadata object
}
\description{
ClusterOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the operation.
}

