% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfig.metadata}
\alias{GceClusterConfig.metadata}
\title{GceClusterConfig.metadata Object}
\usage{
GceClusterConfig.metadata()
}
\value{
GceClusterConfig.metadata object
}
\description{
GceClusterConfig.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Compute Engine metadata entries to add to all instances (see Project and instance metadata (https://cloud.google.com/compute/docs/storing-retrieving-metadata#project_and_instance_metadata)).
}
\seealso{
Other GceClusterConfig functions: \code{\link{GceClusterConfig}}
}
