% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{InstanceGroupConfig}
\alias{InstanceGroupConfig}
\title{InstanceGroupConfig Object}
\usage{
InstanceGroupConfig(numInstances = NULL, instanceNames = NULL,
  imageUri = NULL, machineTypeUri = NULL, diskConfig = NULL,
  isPreemptible = NULL, managedGroupConfig = NULL)
}
\arguments{
\item{numInstances}{The number of VM instances in the instance group}

\item{instanceNames}{The list of instance names}

\item{imageUri}{[Output-only] The Google Compute Engine image resource used for cluster instances}

\item{machineTypeUri}{The Google Compute Engine machine type used for cluster instances}

\item{diskConfig}{Disk option config settings}

\item{isPreemptible}{Specifies that this instance group contains Preemptible Instances}

\item{managedGroupConfig}{[Output-only] The config for Google Compute Engine Instance Group Manager that manages this group}
}
\value{
InstanceGroupConfig object
}
\description{
InstanceGroupConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The config settings for Google Compute Engine resources in an instance group, such as a master or worker group.
}

