% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ListClustersResponse}
\alias{ListClustersResponse}
\title{Google Cloud Dataproc API Objects 
Manages Hadoop-based clusters and jobs on Google Cloud Platform.}
\usage{
ListClustersResponse(clusters = NULL, nextPageToken = NULL)
}
\arguments{
\item{clusters}{Output-only The clusters in the project}

\item{nextPageToken}{Output-only This token is included in the response if there are more results to fetch}
}
\value{
ListClustersResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:42:37
filename: /Users/mark/dev/R/autoGoogleAPI/googledataprocv1.auto/R/dataproc_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListClustersResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The list of all clusters in a project.
}
