% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SoftwareConfig.properties}
\alias{SoftwareConfig.properties}
\title{SoftwareConfig.properties Object}
\usage{
SoftwareConfig.properties()
}
\value{
SoftwareConfig.properties object
}
\description{
SoftwareConfig.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional The properties to set on daemon config files.Property keys are specified in prefix:property format, such as core:fs.defaultFS. The following are supported prefixes and their mappings:core: core-site.xmlhdfs: hdfs-site.xmlmapred: mapred-site.xmlyarn: yarn-site.xmlhive: hive-site.xmlpig: pig.propertiesspark: spark-defaults.conf
}
\seealso{
Other SoftwareConfig functions: \code{\link{SoftwareConfig}}
}
